abc/backend/requirements.txt
uvicorn
fastapi
pydantic
pandas
numpy
scikit-learn
joblib
python-multipart
aiofiles
jinja2
python-dotenv
google.generativeai
sqlalchemy
prophet